+++ Horizen.FullNode.API.Net.Standard/RPCReturnTypes/RPCReturn-Control.cs
@@ -18,17 +18,17 @@ namespace Horizen.FullNode.API.Net.Standard.RPCReturnTypes
         public int version { get; set; }
         public int protocolversion { get; set; }
         public int walletversion { get; set; }
-        public float balance { get; set; }
+        public decimal balance { get; set; }
         public int blocks { get; set; }
         public int timeoffset { get; set; }
         public int connections { get; set; }
         public string proxy { get; set; }
-        public float difficulty { get; set; }
+        public decimal difficulty { get; set; }
         public bool testnet { get; set; }
         public int keypoololdest { get; set; }
         public int keypoolsize { get; set; }
-        public float paytxfee { get; set; }
-        public float relayfee { get; set; }
+        public decimal paytxfee { get; set; }
+        public decimal relayfee { get; set; }
         public string errors { get; set; }
     }
